Trekker ut TranslationFilter i egen komponent, med tilhørende hook
