+++ app/routes/termliste.tsx
@@ -16,7 +16,6 @@ import { data } from '@remix-run/node';
 import { Await, Link, useLoaderData, useRouteLoaderData } from '@remix-run/react';
 import type { ChangeEvent, KeyboardEvent, MouseEvent } from 'react';
 import { Suspense, useState } from 'react';
-import { Form, FormGroup, Input, Label } from 'reactstrap';
 import { ClientOnly } from 'remix-utils/client-only';
 
 import { Spinner } from '~/lib/components/spinner';
@@ -26,34 +25,10 @@ import type { loader as rootLoader } from '~/root';
 import style from '~/styles/termliste.module.css';
 import type { Subject } from '~/types/subject';
 import type { Language, Term } from '~/types/term';
-
-interface TransFilter {
-  text: string;
-  filter: TransFilterType;
-  defaultChecked: boolean;
-}
-
-type TransFilterType = 'all' | 'translated' | 'incomplete';
+import { TranslationFilter } from '~/lib/components/translation-filter';
+import { useTransFilter } from '~/lib/use-trans-filter';
 
 const AllSubjects: Subject = { field: 'Alle fagfelt', subfields: [] };
-
-const transFilters: TransFilter[] = [
-  {
-    text: 'Alle',
-    filter: 'all',
-    defaultChecked: true,
-  },
-  {
-    text: 'Oversatt',
-    filter: 'translated',
-    defaultChecked: false,
-  },
-  {
-    text: 'Ufullstendig',
-    filter: 'incomplete',
-    defaultChecked: false,
-  },
-];
 
 type ServerData = Promise<{
   success: boolean;
@@ -89,23 +64,12 @@ export function loader() {
 export default function Termliste() {
   const { terms } = useRouteLoaderData<typeof rootLoader>('root');
   const subjectsData = useLoaderData<typeof loader>() as unknown as ServerData;
-  const [transFilter, setTransFilter] = useState<TransFilterType>('all');
+  const [setTransFilter, applyTransFilter] = useTransFilter();
   const [subjectFilter, setSubjectFilter] = useState<string | null>(AllSubjects.field);
   const [order, setOrder] = useState<Order>('asc');
   const [orderBy, setOrderBy] = useState<keyof Language>('en');
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
-
-  const applyTransFilter = (terms: Term[]): Term[] => {
-    switch (transFilter) {
-      case 'translated':
-        return terms.filter((term) => term.nb !== '' || term.nn !== '');
-      case 'incomplete':
-        return terms.filter((term) => term.nb === '' || term.nn === '');
-      default:
-        return terms;
-    }
-  };
 
   const applySubjectFilter = (terms: Term[]): Term[] => {
     if (subjectFilter === null) return terms;
@@ -159,21 +123,7 @@ export default function Termliste() {
     <div className="container-sm my-2">
       <div className="col-12 col-lg-10 mx-auto">
         <div className={style.header}>
-          <Form className={style.form}>
-            {transFilters.map((filter) => (
-              <FormGroup check inline key={filter.filter}>
-                <Input
-                  name="dictionaryView"
-                  type="radio"
-                  onChange={() => {
-                    setTransFilter(filter.filter);
-                  }}
-                  defaultChecked={filter.defaultChecked}
-                />
-                <Label check>{filter.text}</Label>
-              </FormGroup>
-            ))}
-          </Form>
+          <TranslationFilter setTransFilter={setTransFilter} />
           {subjectFilterComponent()}
         </div>
         <Paper sx={{ width: '100%', mb: 2, bgcolor: 'background.paper' }}>
