+++ app/styles/termliste-ny.module.css
@@ -1,17 +1,20 @@
 table {
-  border: 1px solid lightgray;
   background: white;
   color: black;
-}
 
-tbody {
-  border-bottom: 1px solid lightgray;
 }
 
 th {
   border-bottom: 1px solid lightgray;
-  border-right: 1px solid lightgray;
-  padding: 2px 4px;
+  padding: 4px 8px;
+}
+
+tr:not(:last-child) {
+  border-bottom: 1px solid lightgray;
+}
+
+td {
+  padding: 4px 8px;
 }
 
 tfoot {
