+++ app/types/filters.ts
@@ -1,0 +1,7 @@
+export type TransFilter = {
+  text: string;
+  filter: TransFilterType;
+  defaultChecked: boolean;
+};
+
+export type TransFilterType = 'all' | 'translated' | 'incomplete';
